+++ login-workflow/src/pages/create-account/steps/provide-email/provide-email.component.ts
@@ -1,10 +1,9 @@
-import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
-import { FormControl, Validators } from '@angular/forms';
+import { Component, EventEmitter, Input, OnInit, Output, ViewChild } from '@angular/core';
+import { FormControl, ValidatorFn } from '@angular/forms';
 
-import { AuthErrorStateMatcher } from '../../../../util/matcher';
 import { PxbAuthConfig } from './../../../../services/config/auth-config';
-
 import { PxbAuthTranslations } from '../../../../translations/auth-translations';
+import { EmailFieldComponent } from '../../../../components/email-field/email-field.component';
 
 @Component({
     selector: 'pxb-create-account-provide-email-step',
@@ -18,29 +17,11 @@ import { PxbAuthTranslations } from '../../../../translations/auth-translations'
         <mat-divider class="pxb-auth-divider" style="margin-top: 16px; margin-bottom: 32px;"></mat-divider>
         <div class="pxb-auth-full-height">
             <form>
-                <mat-form-field appearance="fill" style="width: 100%;">
-                    <mat-label>{{ translate.GENERAL.EMAIL_FORM_LABEL }}</mat-label>
-                    <input
-                        id="email"
-                        name="email"
-                        type="email"
-                        matInput
-                        [formControl]="emailFormControl"
-                        (ngModelChange)="updateEmail(emailFormControl.value)"
-                        [errorStateMatcher]="emailMatcher"
-                        (keyup.enter)="advance.emit(true)"
-                    />
-                    <mat-error
-                        *ngIf="emailFormControl.hasError('email') && !emailFormControl.hasError('required')"
-                        [innerHTML]="translate.GENERAL.EMAIL_INVALID_ERROR"
-                    >
-                    </mat-error>
-                    <mat-error
-                        *ngIf="emailFormControl.hasError('required')"
-                        [innerHTML]="translate.GENERAL.IS_REQUIRED_ERROR(translate.GENERAL.EMAIL_FORM_LABEL)"
-                    >
-                    </mat-error>
-                </mat-form-field>
+                <pxb-email-field
+                    [customEmailValidator]="customEmailValidator"
+                    (enter)="advance.emit()"
+                    (edit)="updateEmail($event)"
+                ></pxb-email-field>
             </form>
         </div>
     `,
@@ -48,12 +29,14 @@ import { PxbAuthTranslations } from '../../../../translations/auth-translations'
 export class PxbProvideEmailComponent implements OnInit {
     @Input() email: string;
     @Input() isValidEmail: boolean;
+    @Input() customEmailValidator: ValidatorFn;
 
     @Output() emailChange: EventEmitter<string> = new EventEmitter<string>();
     @Output() isValidEmailChange: EventEmitter<boolean> = new EventEmitter<boolean>();
-    @Output() advance: EventEmitter<boolean> = new EventEmitter<boolean>();
+    @Output() advance: EventEmitter<void> = new EventEmitter<void>();
 
-    emailMatcher = new AuthErrorStateMatcher();
+    @ViewChild(EmailFieldComponent) emailFieldComponent: EmailFieldComponent;
+
     emailFormControl: FormControl;
     translate: PxbAuthTranslations;
 
@@ -61,12 +44,10 @@ export class PxbProvideEmailComponent implements OnInit {
 
     ngOnInit(): void {
         this.translate = this._pxbAuthConfig.getTranslations();
-        const emailValidators = [
-            Validators.required,
-            Validators.email,
-            Validators.pattern(/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}$/i),
-        ];
-        this.emailFormControl = new FormControl(this.email, emailValidators);
+    }
+
+    ngAfterViewInit(): void {
+        this.emailFormControl = this.emailFieldComponent.emailFormControl;
     }
 
     updateEmail(email: string): void {