+++ login-workflow/src/pages/forgot-password/forgot-password.component.ts
@@ -1,8 +1,7 @@
-import { Component, Input, OnInit } from '@angular/core';
-import { ValidatorFn, FormControl, Validators } from '@angular/forms';
+import { Component, Input, OnInit, ViewChild } from '@angular/core';
+import { ValidatorFn, FormControl } from '@angular/forms';
 import { Router } from '@angular/router';
 import { AuthErrorStateMatcher } from '../../util/matcher';
-import { PXB_LOGIN_VALIDATOR_ERROR_NAME } from '../login/login.component';
 import { PxbAuthUIService } from '../../services/api/auth-ui.service';
 import { AUTH_ROUTES } from '../../auth/auth.routes';
 import { PxbAuthConfig } from '../../services/config/auth-config';
@@ -10,6 +9,7 @@ import { PxbAuthSecurityService } from '../../services/state/auth-security.servi
 import { PxbForgotPasswordErrorDialogService } from '../../services/dialog/forgot-password-error-dialog.service';
 import { ErrorDialogData } from '../../services/dialog/error-dialog.service';
 import { PxbAuthTranslations } from '../../translations/auth-translations';
+import { EmailFieldComponent } from '../../components/email-field/email-field.component';
 
 @Component({
     selector: 'pxb-forgot-password',
@@ -18,9 +18,9 @@ import { PxbAuthTranslations } from '../../translations/auth-translations';
 })
 export class PxbForgotPasswordComponent implements OnInit {
     @Input() customEmailValidator: ValidatorFn;
+    @ViewChild(EmailFieldComponent) emailFieldComponent: EmailFieldComponent;
 
-    customErrorName = PXB_LOGIN_VALIDATOR_ERROR_NAME;
-
+    email: string;
     emailFormControl: FormControl;
     matcher = new AuthErrorStateMatcher();
     passwordResetSuccess = false;
@@ -37,15 +37,10 @@ export class PxbForgotPasswordComponent implements OnInit {
 
     ngOnInit(): void {
         this.translate = this._pxbAuthConfig.getTranslations();
-        const emailValidators = [
-            Validators.required,
-            Validators.email,
-            Validators.pattern(/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}$/i),
-        ];
-        if (this.customEmailValidator) {
-            emailValidators.push(this.customEmailValidator);
-        }
-        this.emailFormControl = new FormControl('', emailValidators);
+    }
+
+    ngAfterViewInit(): void {
+        this.emailFormControl = this.emailFieldComponent.emailFormControl;
     }
 
     navigateToLogin(): void {
@@ -57,11 +52,14 @@ export class PxbForgotPasswordComponent implements OnInit {
         return this.translate.FORGOT_PASSWORD.CONTACT_SUPPORT_BY_PHONE(phone);
     }
 
+    hasValidEmail(): boolean {
+        return this.emailFormControl && this.emailFormControl.valid;
+    }
+
     resetPassword(): void {
-        const email = this.emailFormControl.value;
         this._pxbSecurityService.setLoading(true);
         this._pxbAuthUIActionsService
-            .forgotPassword(email)
+            .forgotPassword(this.email)
             .then(() => {
                 this.passwordResetSuccess = true;
                 void this._router.navigate([`${AUTH_ROUTES.AUTH_WORKFLOW}/${AUTH_ROUTES.FORGOT_PASSWORD}`]);