+++ login-workflow/CHANGELOG.md
@@ -5,11 +5,18 @@
 ### Added
 
 -  Added ability to suppress optional pages (Contact Support, Create Account, etc) via `PxbAuthConfig`. 
+-  Added `customEmailValidator` prop to `PxbAuthConfig`.
+-  Added `inferOnAuthenticatedRoute` to `PxbAuthSecurityService`.
+-  Added text and content overrides via `PxbAuthTranslations`.
 
 ### Changed
 
 -  Changed how custom account details can be provided during account registration.
--  Removed ability to skip first and last name during account registration. 
+-  Changed how custom password requirements are configured in `PxbAuthConfig`.
+
+### Removed
+
+-  Removed ability to skip first & last name during account registration.  
 
 ## v1.0.1